chore: factor out workspace settings from task
