+++ src/main/rpc.ts
@@ -32,6 +32,7 @@ import { telemetryController } from './core/telemetry/controller';
 import { terminalsController } from './core/terminals/controller';
 import { updateController } from './core/updates/controller';
 import { viewStateController } from './core/view-state/controller';
+import { projectSettingsController } from './core/workspaces/project-settings-controller';
 import { workspaceController } from './core/workspaces/controller';
 import { legacyPortController } from './db/legacy-port/controller';
 
@@ -71,6 +72,7 @@ export const rpcRouter = createRPCRouter({
   viewState: viewStateController,
   search: searchController,
   workspaces: workspaceController,
+  projectSettings: projectSettingsController,
 });
 
 export type RpcRouter = typeof rpcRouter;
